Moved poll + select to net_calls.rs
